feat(spec): support full spec (#152)
* fix(spacejam-fuzz): set self.init after init_state in Fuzzer::handle_single

* fix(runtime): authorization queue codec and α' transition

* fix: guard arithmetic overflow sites surfaced by fuzz

* fix(crypto): align merkle::trie encoding with GP spec

* feat(spec): gate constants on tiny/full features

* feat(codec): split Array into FixedArray and heap-backed Array for full-spec support

* feat(spec): add full-spec feature gates for erasure, crypto, and testing

* fix(runtime): enforce strict preimage ordering and prune clones

* perf(crypto): batch ed25519 verify with rayon par_chunks

* fix(vm): preserve validators when accumulating service with no code

* chore(logs): demote expected-during-fuzz warnings to debug, default ansi off

* style: cargo fmt

* chore(clippy): make clippy happy

* chore: fmt

+++ crates/offchain/worker/src/assurer.rs
@@ -117,7 +117,7 @@ pub trait Assurer: DataLake {
                 None
             };
 
-            let mut segment = [0u8; score::SEGMENT_SIZE as usize];
+            let mut segment = [0u8; score::SEGMENT_SIZE];
             segment.copy_from_slice(&segment_shard);
             results.push((segment, justification));
         }
